Fixed Checkbox Table width & height issues
This update fixes the problem with the width and height of the table in which the checkboxes of the player are in. The issue happened due to the `input[type=checkbox]` being given a property of `Visibility: hidden` instead of `Display: none`. Due to the `Input[type=checkbox]` still occupying space, it would affect the width and height of the element it was in.