Implemented Theme-able Scrollbar for Song List
The default scrollbar in browsers was hidden and was replaced by a "scrollbar" created separately. The new scrollbar, just like the media player's song track uses Input[type=range] to make it seekable and onmousedown, onmouseclick, onmouseup functions for the "scrollbar"'s up and down buttons. The end result is a scrollbar that functions as you'd expect.

Since the system's default scrollbar was hidden, this also fixes the issue present on Google Chrome and Opera when the scrollbar glitches out when using it.

However, this brings a new issue. The media player's theme uses CSS3's filter property to increase the birhgtness of certain elements when on hover; this implementation saves a lot of code since the color of elements do not have to be set separately on a per-theme basis to have the same effect. The bad thing about this is that filter seems to glitch out the new scrollbar when used in the song list which makes it flicker when hovering over a song title.